chore: remove redundant create rollup util
