+++ src/createRollupPrepareTransactionRequest.ts
@@ -1,6 +1,6 @@
 import { Address, PublicClient, encodeFunctionData, zeroAddress } from 'viem';
 
-import { CreateRollupFunctionInputs, CreateRollupParams } from './createRollup';
+import { CreateRollupFunctionInputs, CreateRollupParams } from './createRollupTypes';
 import { defaults } from './createRollupDefaults';
 import { createRollupGetCallValue } from './createRollupGetCallValue';
 import { createRollupGetMaxDataSize } from './createRollupGetMaxDataSize';
